% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lambda.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{Calculate lambda}
\usage{
calculate_lambda(stat, df)
}
\arguments{
\item{stat}{A vector of p-values}

\item{df}{Degrees of freedom}
}
\description{
Calculate the genomic inflation factor from a vector of p-values
}
