% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dosage_est.R
\name{extract_dosage_est}
\alias{extract_dosage_est}
\title{Extract alternate allele dosage estimates from a GDS}
\usage{
extract_dosage_est(gds, chrom, from.bp, to.bp)
}
\arguments{
\item{gds}{A GDS connection}

\item{chrom}{A vector of chromosome numbers}

\item{from.bp}{A vector of starting BP positions}

\item{to.bp}{A vector of ending BP positions}
}
\description{
Extract alternate allele dosage estimates from a GDS
}
