% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_resids.R
\name{make_resids}
\alias{make_resids}
\title{Run survival analysis and compute residuals on a phenotype dataframe}
\usage{
make_resids(
  x,
  age,
  status,
  surv_name = "surv",
  covars,
  keep_reg = FALSE,
  mgl_name = "martingale",
  dev_name = "devres",
  weighted = FALSE
)
}
\arguments{
\item{x}{A dataframe with age, status and covariate variables}

\item{age}{The name of the age variable}

\item{status}{The name of the event variable}

\item{surv_name}{The name of the survival variable in the returned dataframe}

\item{covars}{The names of the covariates for the Cox regression}

\item{keep_reg}{Indicator for the function to return the Cox regression object}

\item{mgl_name}{The name of the Martingale residual variable in the returned dataframe}

\item{dev_name}{The name of the deviance residual variable in the returned dataframe}

\item{weighted}{Passed as the argument \code{weighted} to \code{residuals()}}
}
\description{
Run survival analysis and compute residuals on a phenotype dataframe
}
