% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sexcheck.R
\name{read_sexcheck}
\alias{read_sexcheck}
\title{Read plink.sexcheck}
\usage{
read_sexcheck(plink.sexcheck, ...)
}
\arguments{
\item{plink.sexcheck}{Output from \code{plink --check-sex}.}

\item{...}{Arguments passed throught to read.table().}
}
\description{
Read in the output of \code{plink --check-sex} as a dataframe.
}
