% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_var_by_chrom.R
\name{subset_var_by_chrom}
\alias{subset_var_by_chrom}
\title{Subset a vector of variant IDs to the specified chromosome(s)}
\usage{
subset_var_by_chrom(variant_id, chrom, gds)
}
\arguments{
\item{variant_id}{A vector of variant IDs}

\item{chrom}{Chromosome number or numbers}

\item{gds}{a GDS connection}
}
\description{
Subset a vector of variant IDs to the specified chromosome(s)
}
